+++ cli/run.ts
@@ -1,3 +1,4 @@
+import { resolveConfigFile } from "../core/utils/lume_config.ts";
 import { createSite } from "./utils.ts";
 
 /** Run one or more custom scripts */
@@ -5,7 +6,8 @@ export async function run(
   config: string | undefined,
   scripts: string[],
 ) {
-  const site = await createSite(config);
+  const _config = await resolveConfigFile(["_config.ts", "_config.js"], config);
+  const site = await createSite(_config);
 
   for (const script of scripts) {
     const success = await site.run(script);