+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 - Added missing tests for SEO plugin.
 - Renamed `Config` interface of `seo` and `validate_html` plugins to `Options`.
 - Added export-ignore paths to `.gitattributes`.
+- `site.remote()`: support commit SHA
 - Updated dependencies: `deno-loader`, `tailwindcss`, `html-validate`, `magic-string`, `minify-html` and some icons.
 
 ## [3.1.1] - 2025-10-20